[OC-939] Synchronize timeline with feed
