+++ ui/main/src/app/modules/feed/components/time-line/init-chart/init-chart.component.html
@@ -42,9 +42,7 @@
    (zoomChange) call when custom timeline chart change his domain
   -->
   <of-custom-timeline-chart class="of-custom-timeline-chart"
-                              *ngIf="cardsData"
                              [prod]="true"
-                             [cardsData]="cardsData"
                              [domainId]="domainId"
                              [valueDomain]="myDomain"
                              [followClockTick]="followClockTick"
