+++ ui/main/src/app/modules/feed/components/time-line/init-chart/init-chart.component.ts
@@ -2,13 +2,12 @@
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import * as _ from 'lodash';
 import * as moment from 'moment';
-import { Observable, of, Subscription } from 'rxjs';
-import { select, Store } from '@ngrx/store';
+import {  Subscription } from 'rxjs';
+import { Store } from '@ngrx/store';
 import { AppState } from '@ofStore/index';
-import * as timelineSelectors from '@ofSelectors/timeline.selectors';
-import { catchError, debounceTime, distinctUntilChanged } from 'rxjs/operators';
 import { FilterType } from '@ofServices/filter.service';
 import { ApplyFilter } from '@ofActions/feed.actions';
+
 
 const forwardWeekConf = {
   start: { year: 0, month: 0, week: 1, day: 0, hour: 0, minute: 0, second: 0 },
@@ -28,7 +27,6 @@ export class InitChartComponent implements OnInit, OnDestroy {
   public cardsData: any[];
   public myDomain: number[];
   public domainId: string;
-  data$: Observable<any[]>;
   subscription: Subscription;
 
   // required for domain movements specifications
@@ -52,7 +50,6 @@ export class InitChartComponent implements OnInit, OnDestroy {
    */
   ngOnInit() {
     this.initDomains();
-    this.initDataPipe();
   }
 
   /**
@@ -177,16 +174,6 @@ export class InitChartComponent implements OnInit, OnDestroy {
     }));
   }
 
-
-  initDataPipe(): void {
-    this.data$ = this.store.pipe(
-      select(timelineSelectors.selectTimelineSelection),
-      catchError(err => of([]))
-    );
-    this.subscription = this.data$.pipe(debounceTime(300), distinctUntilChanged()).subscribe(value => {
-      this.cardsData  = _.cloneDeep(value);
-    });
-  }
 
   /**
  * unsubscribe every subscription made on this file
