--- ui/main/src/app/modules/feed/components/time-line/time-line.component.spec.ts
@@ -1,145 +1,0 @@
-
-
-import {async, ComponentFixture, TestBed} from '@angular/core/testing';
-import {TimeLineComponent} from './time-line.component';
-import {APP_BASE_HREF, CommonModule} from '@angular/common';
-import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {FormsModule} from '@angular/forms';
-import {NgxChartsModule} from '@swimlane/ngx-charts';
-import {Store, StoreModule} from '@ngrx/store';
-import {appReducer, AppState, storeConfig} from '@ofStore/index';
-import {RouterTestingModule} from '@angular/router/testing';
-import {RouterStateSerializer, StoreRouterConnectingModule} from '@ngrx/router-store';
-import {CustomTimelineChartComponent} from './custom-timeline-chart/custom-timeline-chart.component';
-import {InitChartComponent} from './init-chart/init-chart.component';
-import {CustomRouterStateSerializer} from '@ofStates/router.state';
-import {NO_ERRORS_SCHEMA} from '@angular/core';
-import {getOneRandomLightCard} from '@tests/helpers';
-import {LoadLightCardsSuccess} from '@ofActions/light-card.actions';
-import {LightCard} from '@ofModel/light-card.model';
-import * as fromStore from '@ofSelectors/feed.selectors';
-import * as timelineSelectors from '@ofSelectors/timeline.selectors';
-import {MouseWheelDirective} from './directives/mouse-wheel.directive';
-import {debounceTime, distinctUntilChanged} from 'rxjs/operators';
-import {compareBySeverityPublishDate} from '@ofStates/feed.state';
-import {TimeService} from '@ofServices/time.service';
-
-describe('TimeLineComponent', () => {
-  let component: TimeLineComponent;
-  let store: Store<AppState>;
-  let fixture: ComponentFixture<TimeLineComponent>;
-;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      imports: [CommonModule,
-        BrowserAnimationsModule,
-        FormsModule,
-        StoreModule.forRoot(appReducer, storeConfig),
-        RouterTestingModule,
-        StoreRouterConnectingModule,
-        NgxChartsModule ],
-      declarations: [ TimeLineComponent, CustomTimelineChartComponent, InitChartComponent,MouseWheelDirective],
-
-      providers: [{provide: APP_BASE_HREF, useValue: '/'},
-        {provide: Store, useClass: Store},
-        {provide: RouterStateSerializer, useClass: CustomRouterStateSerializer},
-        {provide: TimeService, useClass: TimeService}],
-      schemas: [ NO_ERRORS_SCHEMA ]
-    })
-    .compileComponents();
-    store = TestBed.get(Store);
-    spyOn(store, 'dispatch').and.callThrough();
-    fixture = TestBed.createComponent(TimeLineComponent);
-    component = fixture.componentInstance;
-  }));
-
-  it('should init timeline', () => {
-    component.ngOnInit();
-    fixture.detectChanges();
-    expect(component).toBeTruthy();
-  });
-
-
-  it('should create a list with one element when there is ' +
-    'only one card in the state', (done) => {
-    fixture.detectChanges();
-    const oneCard = getOneRandomLightCard();
-    const action = new LoadLightCardsSuccess({lightCards: [oneCard] as LightCard[]});
-    store.dispatch(action);
-    const lightCards$ = store.select(fromStore.selectFeed);
-    lightCards$.pipe(debounceTime(300), distinctUntilChanged())
-        .subscribe(lightCard => {
-      expect(lightCard).toEqual([oneCard]);
-    });
-    const dataCard = [{
-      date: oneCard.startDate,
-      severity: oneCard.severity,
-      publisher: oneCard.publisher,
-      publisherVersion: oneCard.publisherVersion,
-      summary: oneCard.title
-    }];
-    const data$ = store.select((timelineSelectors.selectTimelineSelection));
-    data$.pipe(debounceTime(300), distinctUntilChanged())
-        .subscribe(value => {
-          expect(value).toEqual(dataCard);
-          done();
-        });
-    expect(store.dispatch).toHaveBeenCalledWith(action);
-    expect(component).toBeTruthy();
-
-  });
-
-  it('should create four different circles when there is ' +
-      'four cards with different severity in the state', (done) => {
-      const today = new Date().getTime();
-    fixture.detectChanges();
-    const compliantCard = getOneRandomLightCard({severity: 'COMPLIANT', publishDate: today});
-    const actionCard = getOneRandomLightCard({severity: 'ACTION', publishDate: today});
-    const alarmCard = getOneRandomLightCard({severity: 'ALARM', publishDate: today});
-    const informationCard = getOneRandomLightCard({severity: 'INFORMATION', publishDate: today});
-    const action = new LoadLightCardsSuccess({lightCards: [alarmCard,actionCard,compliantCard,informationCard] as LightCard[]});
-    store.dispatch(action);
-    const lightCards$ = store.select(fromStore.selectSortedFilteredLightCards);
-    lightCards$.pipe(debounceTime(300), distinctUntilChanged())
-        .subscribe(lightCards => {
-      expect(lightCards).toEqual([informationCard, alarmCard, actionCard, compliantCard].sort(compareBySeverityPublishDate)); //Default sort
-    });
-    const dataCard = [{
-        date: alarmCard.startDate,
-        severity: alarmCard.severity,
-        publisher: alarmCard.publisher,
-        publisherVersion: alarmCard.publisherVersion,
-        summary: alarmCard.title
-    }, {
-        date: actionCard.startDate,
-        severity: actionCard.severity,
-        publisher: actionCard.publisher,
-        publisherVersion: actionCard.publisherVersion,
-        summary: actionCard.title
-    }, {
-        date: compliantCard.startDate,
-        severity: compliantCard.severity,
-        publisher: compliantCard.publisher,
-        publisherVersion: compliantCard.publisherVersion,
-        summary: compliantCard.title
-    }, {
-        date: informationCard.startDate,
-        severity: informationCard.severity,
-        publisher: informationCard.publisher,
-        publisherVersion: informationCard.publisherVersion,
-        summary: informationCard.title
-    }];
-    const data$ = store.select((timelineSelectors.selectTimelineSelection));
-    data$.pipe(debounceTime(300), distinctUntilChanged())
-        .subscribe(value => {
-          expect(value).toEqual(dataCard);
-          done();
-        });
-    expect(store.dispatch).toHaveBeenCalledWith(action);
-    expect(component).toBeTruthy();
-
-  });
-
-
-});
