+++ ui/main/src/app/modules/feed/components/time-line/time-line.component.ts
@@ -1,14 +1,11 @@
 
 
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { Observable, of, Subscription } from 'rxjs';
-import { LightCard } from '@ofModel/light-card.model';
+import { of, Subscription } from 'rxjs';
 import { select, Store } from '@ngrx/store';
-import { catchError, debounceTime, distinctUntilChanged } from 'rxjs/operators';
+import { catchError} from 'rxjs/operators';
 import { AppState } from '@ofStore/index';
-import { SetCardDataTimeline } from '@ofActions/timeline.actions';
 import * as _ from 'lodash';
-import * as feedSelectors from '@ofSelectors/feed.selectors';
 import { buildConfigSelector } from '@ofStore/selectors/config.selectors';
 import { buildSettingsOrConfigSelector } from '@ofStore/selectors/settings.x.config.selectors';
 import * as moment from 'moment';
@@ -18,8 +15,7 @@ import * as moment from 'moment';
     templateUrl: './time-line.component.html',
 })
 export class TimeLineComponent implements OnInit, OnDestroy {
-    lightCards$: Observable<LightCard[]>;
-    subscription: Subscription;
+
     localSubscription: Subscription;
 
     public confDomain = [];
@@ -36,9 +32,6 @@ export class TimeLineComponent implements OnInit, OnDestroy {
             l => moment.locale(l)
         )
 
-        this.subscription = this.store.pipe(select(feedSelectors.selectFeed))
-            .pipe(debounceTime(300), distinctUntilChanged())
-            .subscribe(value => this.sendAllCardsToDrawOnTheTimeLine(value));
     }
 
     loadConfiguration() {
@@ -85,34 +78,9 @@ export class TimeLineComponent implements OnInit, OnDestroy {
         });
     }
 
-    sendAllCardsToDrawOnTheTimeLine(cards) {
-        const myCardsTimeline = [];
-        for (const card of cards) {
-            if (card.timeSpans && card.timeSpans.length > 0) {
-                card.timeSpans.forEach(timeSpan => {
-                    const myCardTimelineTimespans = {
-                        date: timeSpan.start, 
-                        severity: card.severity, publisher: card.publisher,
-                        publisherVersion: card.publisherVersion, summary: card.title
-                    };
-                    myCardsTimeline.push(myCardTimelineTimespans);
-                });
-            } else {
-                const myCardTimeline = {
-                    date: card.startDate,
-                    severity: card.severity, publisher: card.publisher,
-                    publisherVersion: card.publisherVersion, summary: card.title
-                };
-                myCardsTimeline.push(myCardTimeline);
-            }
-        }
-        this.store.dispatch(new SetCardDataTimeline({ cardsTimeline: myCardsTimeline }));
-    }
 
     ngOnDestroy() {
-        if (this.subscription) {
-            this.subscription.unsubscribe();
-        }
+
         if (this.localSubscription) {
             this.localSubscription.unsubscribe();
         }
