--- ui/main/src/app/store/actions/timeline.actions.ts
@@ -1,14 +1,0 @@
-
-
-import {Action} from '@ngrx/store';
-
-export enum TimelineActionTypes {
-    SetCardDataTimeline = '[timeline] Set Card Data'
-}
-
-export class SetCardDataTimeline implements Action {
-    readonly type = TimelineActionTypes.SetCardDataTimeline
-    constructor(public payload:{cardsTimeline: any}){}
-}
-
-export type TimelineActions = SetCardDataTimeline;
