+++ ui/main/src/app/store/index.ts
@@ -12,7 +12,6 @@ import {CardOperationEffects} from '@ofEffects/card-operation.effects';
 import {AuthenticationEffects} from '@ofEffects/authentication.effects';
 import {RouterEffects} from 'ngrx-router';
 import {CardFeedState} from '@ofStates/feed.state';
-import {reducer as timelineReducer} from '@ofStore/reducers/timeline.reducer';
 import {reducer as lightCardReducer} from '@ofStore/reducers/light-card.reducer';
 import {reducer as cardReducer} from '@ofStore/reducers/card.reducer';
 import {reducer as configReducer} from '@ofStore/reducers/config.reducer';
@@ -28,7 +27,6 @@ import {LightCardEffects} from '@ofEffects/light-card.effects';
 import {FeedFiltersEffects} from '@ofEffects/feed-filters.effects';
 import {ConfigState} from '@ofStates/config.state';
 import {ConfigEffects} from '@ofEffects/config.effects';
-import {TimelineState} from '@ofStates/timeline.state';
 import {SettingsState} from '@ofStates/settings.state';
 import {SettingsEffects} from '@ofEffects/settings.effects';
 import {ArchiveState} from '@ofStates/archive.state';
@@ -44,7 +42,6 @@ import {TranslateEffects} from '@ofEffects/translate.effects';
 export interface AppState {
     router: RouterReducerState<RouterStateUrl>;
     feed: CardFeedState;
-    timeline: TimelineState;
     authentication: AuthState;
     card: CardState;
     menu: MenuState;
@@ -73,7 +70,6 @@ export const appEffects = [
 export const appReducer: ActionReducerMap<AppState> = {
     router: fromRouter.routerReducer,
     feed: lightCardReducer,
-    timeline: timelineReducer,
     authentication: authenticationReducer,
     card: cardReducer,
     menu: menuReducer,
