--- ui/main/src/app/store/selectors/timeline.selectors.ts
@@ -1,10 +1,0 @@
-
-
-import {createSelector} from '@ngrx/store';
-import {AppState} from "@ofStore/index";
-
-export const selectTimelineState = (state: AppState) => state.timeline;
-
-export const selectTimelineSelection = createSelector(selectTimelineState,
-    state => state.data);
-
